Create example globals.php
Since the variables in globals.php can be different depending on how
LEAF is deployed, each globals.php was added to the .gitignore to
prevent accidentally committing those setup specific variables.

This should also prevent the need to stash or copy over changes in
these files when LEAF is being upgraded in production.

New setup instructions were added to the README.