Added UI support for Digital Signing in Workflows
Added option to require Digital Signature on a step in the Workflow
Editor.

Added description text on form that requires a Digital Signature which
explains necessary actions for the User to take.

Added Workflow Javascript API.

Added endpoint to WorkflowController to set if a Workflow Step requires
a Digital Signature.