Dev (#233)
* Update file extension whitelists (#230)

Updating all file extension whitelists to include support for .ai and
.eps files.

* Allow "Unassigned" employee selector (#231)

If a user selects an Employee from the employee selector for a form
indicator value, previously they were not allowed to set that indicator
as "Unassigned". The selector would persist the previous value. Now a
user can clear out that indicator data.