Update jsonExport_PDL.php - Fortify
the htmlspecialchars function is used to escape any special characters that might be included in the JSON output. The ENT_QUOTES flag is used to escape both single and double quotes, and the UTF-8 charset is used to encode the output. Finally, the JSON is outputted using the echo statement. This was requested by the fortify team.