LEAF-4165 - User Access Group History
### Summary:

This is a Help Desk ticket issue that was brought up by a user who was not able to see the deleted timestamp in the history modal.

### Potential Impact:

The user will be able to see the deleted user from the group and pruned from the group as well.

### Testing:

-Removed:

1. Admin > User Access > Select User Group

2. Add Employee > Select User > Click "Save"

1. Click "Delete" > Click "View History"

You should be able to see "user" removed user: "removed user"

- Prune:

1. Admin > User Access > Select User Group

2. Click "Show Inactive Users" >

3. Click "Prune" > Click "View History"

You should be able to see "user" pruned user: "removed user"